Spiffy things up
Add coverage and linting. Make the code and example Standard
compliant. Put a concrete assertion in the example.

Refactor the JS, use more modern class syntax, remove duplicated code.

I'm not 100% thrilled with passing in the numeric selector to the
process() method. It does seem more sane than passing in lambdas
and/or turning the function completely inside-out.