More refactoring
Remove the odd array-selection syntax and instead use the select
argument to choose the position of the alphabets to be operated
on. This is slightly more clear and a good bit shorter.

Put rotate on String directly. Remove the modulus, it'll never
actually do anything (as this is only called on the string the length
is derived from if it's not 1). If rotateAmount is zero, do the
rotation anyways to save a line of code.

Get rid of the charAt/indexOf methods and have the caller just use the
ones on data directly. Exchange 10 added characters for 8 removed
lines. Demeter be damned.